chore: moved put request to [user].ts
