+++ src/app/api/user/[userId].ts
@@ -1,6 +1,6 @@
 import connectDB from "@database/db";
 import { NextResponse } from "next/server";
-import User from "@database/userSchema";
+import User, { IUser } from "@database/userSchema";
 
 // get endpoint to get indivdual user
 /**
@@ -20,5 +20,21 @@ export async function GET(id: string) {
   } catch (error) {
     console.error("Error fetching user", error);
     return new NextResponse("Error fetching user", { status: 500 });
+  }
+}
+
+
+// put endpoint to upload user to mongoDB
+/**
+ * @param {object} data - user object to put in db
+ * @return {object} - object that was uploaded to db
+ */
+export async function PUT(data: IUser) {
+  try {
+    await connectDB();
+    const newUser = await User.create(data);
+    return NextResponse.json({ newUser });
+  } catch (error) {
+    console.error("error adding user", error);
   }
 }
