+++ src/app/api/user/users.ts
@@ -1,6 +1,6 @@
 import connectDB from "@database/db";
 import { NextResponse } from "next/server";
-import User, { IUser } from "@database/userSchema";
+import User from "@database/userSchema";
 
 // get endpoint to return array of all users
 /**
@@ -13,20 +13,5 @@ export async function GET() {
     return NextResponse.json({ allUsers });
   } catch (error) {
     console.error("error fetching the users", error);
-  }
-}
-
-// put endpoint to upload user to mongoDB
-/**
- * @param {object} data - user object to put in db
- * @return {object} - object that was uploaded to db
- */
-export async function PUT(data: IUser) {
-  try {
-    await connectDB();
-    const newUser = await User.create(data);
-    return NextResponse.json({ newUser });
-  } catch (error) {
-    console.error("error adding user", error);
   }
 }
